refactor: extract grant display into _display/_grants.py
Move _grant_description and helpers out of _race.py for reuse by
background screens without coupling grant rendering to race module.
